Fetch the source and target refs for a PR

Fetch the source and target refs for a PR

Usage

FetchPRRefs(
  intPRNumber = FetchLatestRepoPRNumber(strOwner, strRepo, strGHToken, "open"),
  strOwner = gh::gh_tree_remote()[["username"]],
  strRepo = gh::gh_tree_remote()[["repo"]],
  strGHToken = gh::gh_token(),
  envCall = rlang::caller_env()
)

Arguments

intPRNumber

(length-1 integer) The number of the pull request to fetch information about.

strOwner

(length-1 character) GitHub username or organization name.

strRepo

(length-1 character) GitHub repository name.

strGHToken

(length-1 character) GitHub token with permissions to read issues.

envCall

(environment) The environment to use for error reporting. Typically set to rlang::caller_env().

Value

A named character vector with strSourceRef and strTargetRef.