Printing qcthat objects

One of the main purposes of qcthat is to produce user-friendly reports about a package's issue and test statuses. Printing an issue-test matrix provides a nested view of the relationship between milestones, issues, and tests, as well as the state of issues ("open", "closed (completed)", or "closed (won't fix)") and the disposition of tests ("passed", "failed", or "skipped").

Usage

# S3 method for class 'qcthat_Object'
print(x, ...)

# S3 method for class 'qcthat_Object'
format(
  x,
  ...,
  lglUseEmoji = getOption("qcthat.emoji", TRUE),
  lglShowMilestones = TRUE,
  lglShowIgnoredLabels = TRUE,
  fnTransform = identity
)

Arguments

x

(qcthat_Object) The qcthat object to format.

...

Additional arguments passed to methods.

lglUseEmoji

(length-1 logical) Whether to use emojis (if TRUE and the emoji package is installed) or ASCII indicators (if FALSE) in the output. By default, this is determined by the qcthat.emoji option, which defaults to TRUE.

lglShowMilestones

(length-1 logical) Whether to separate issues by milestones in reports.

lglShowIgnoredLabels

(length-1 logical) Whether to show information in reports about issue labels (such as "qcthat-nocov") that have been ignored.

fnTransform

(function) A function to transform the output before returning it. By default, this is base::identity() for format() (returns the formatted output), and base::writeLines() for print() (sends the output to stdout()).

Examples

if (FALSE) { # interactive()

  lTestResults <- testthat::test_local(
    stop_on_failure = FALSE,
    reporter = "silent"
  )
  IssueTestMatrix <- CompileIssueTestMatrix(
    dfRepoIssues = FetchRepoIssues(),
    dfTestResults = CompileTestResults(lTestResults)
  )
  print(IssueTestMatrix)

  # Remove the "qcthat_IssueTestMatrix" class to see the raw tibble.
  print(tibble::as_tibble(IssueTestMatrix))
}