Generate a QC report of a specific milestone or milestones

Generate a report about the test status of issues in a milestone or milestones.

Usage

QCMilestones(
  chrMilestones,
  strPkgRoot = ".",
  strOwner = GetGHOwner(strPkgRoot),
  strRepo = GetGHRepo(strPkgRoot),
  strGHToken = gh::gh_token(),
  lglWarn = TRUE,
  chrIgnoredLabels = DefaultIgnoreLabels(),
  envCall = rlang::caller_env()
)

Arguments

chrMilestones

(character) The name(s) of milestone(s) to filter issues by.

strPkgRoot

(length-1 character) The path to a directory in the package. Will be expanded using gert::git_find().

strOwner

(length-1 character) GitHub username or organization name.

strRepo

(length-1 character) GitHub repository name.

strGHToken

(length-1 character) GitHub token with permissions to read issues.

lglWarn

(length-1 logical) Whether to warn when an extra value is included in the filter (but the report still returns results). Defaults to TRUE.

chrIgnoredLabels

(character) GitHub labels to ignore, such as "qcthat-nocov".

envCall

(environment) The environment to use for error reporting. Typically set to rlang::caller_env().

Value

A qcthat_IssueTestMatrix object as returned by QCPackage(), filtered to the indicated issues.

Examples

if (FALSE) { # interactive()

  # This will only make sense if you are working in a git repository that has
  # a milestone named "v0.1.0" on GitHub.
  QCMilestones("v0.1.0")
}